requirements.txt
pytest
pytest-selenium
termcolor
allure-python-commons
python-dotenv